requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
psycopg[binary]
python-multipart==0.0.9
